drivers: sensor: apds9960: move register definitions to private header
Drop all the symbols that were incorrectly part of the public header
and move them to a driver-internal header.
